+++ agent/src/metric/document.rs
@@ -292,7 +292,7 @@ pub struct Tagger {
     pub protocol: IpProtocol,
     pub acl_gid: u16,
     pub server_port: u16, // tunnel_ip_id also uses this field
-    pub vtap_id: u16,
+    pub vtap_id: u32,
     pub tap_port: TapPort,
     pub tap_type: TapType,
     pub l7_protocol: L7Protocol,
@@ -384,7 +384,7 @@ impl From<Tagger> for metric::MiniTag {
                 protocol: u8::from(t.protocol) as u32,
                 acl_gid: t.acl_gid as u32,
                 server_port: t.server_port as u32,
-                vtap_id: t.vtap_id as u32,
+                vtap_id: t.vtap_id,
                 tap_port: t.tap_port.0,
                 tap_type: u16::from(t.tap_type) as u32,
                 l7_protocol: t.l7_protocol as u32,
