+++ agent/src/platform/prometheus/targets.rs
@@ -255,7 +255,7 @@ impl TargetsWatcher {
             PrometheusApiSyncRequest {
                 cluster_id: Some(config_guard.kubernetes_cluster_id.to_string()),
                 version: pb_version,
-                vtap_id: Some(config_guard.vtap_id as u32),
+                vtap_id: Some(config_guard.vtap_id),
                 source_ip: Some(id.ip.to_string()),
                 team_id: Some(id.team_id.clone()),
                 error_msg: Some(err_msgs.join(";")),
